fix: git 충돌 수동 merge -  폴더 이름이 바뀌어서 충돌이 나지 않았음 -  수동으로 충돌 해결
