+++ client/src/components/Lobby/index.js
@@ -1,17 +1,47 @@
 import React, { useState, useEffect } from 'react';
+import { useHistory } from 'react-router-dom';
 import jwt from 'jsonwebtoken';
 import {} from 'dotenv/config';
 import cookie from 'cookie';
 import socket from '../../modules/socket';
+import RoomInfoButton from './RoomInfoButton';
+import GitHubLoginButton from './GitHubLoginButton';
+import RoomCreateModal from './RoomCreateModal';
 import {
-  LobbyWrapper, LobbyHeader, LobbyBody, LobbyNickname, CreateRoomButton, RoomInfoButton,
+  LobbyWrapper, LobbyHeader, LobbyBody, LobbyNickname, CreateRoomButton,
 } from './style';
 
 const privateKey = process.env.REACT_APP_JWT_SECRET_KEY;
 const algorithm = process.env.REACT_APP_JWT_ALGORITHM;
+// 아래는 서버 연결 전 더미 데이터임.
+const dummyRoomInfos = [
+  {
+    id: '1', name: '보현님 언능 들어오세요', numOfUsers: 2, isEnterable: true,
+  },
+  {
+    id: '2', name: '형규님 그만 주무세요', numOfUsers: 2, isEnterable: false,
+  },
+  {
+    id: '3', name: '희선님... 말 잘들을게요 충성^^7', numOfUsers: 2, isEnterable: true,
+  },
+];
 
 const Lobby = () => {
   const [userName, setUserName] = useState('guest');
+  const [isModalOpen, setModalOpen] = useState(false);
+  const history = useHistory();
+
+  const RoomInfoButtons = () => dummyRoomInfos.map(({
+    id, name, numOfUsers, isEnterable,
+  }) => (
+    <RoomInfoButton
+      key={id}
+      roomId={id}
+      name={name}
+      numOfUsers={numOfUsers}
+      enterable={isEnterable} />
+  ));
+  const openRoomCreateModal = () => setModalOpen(true);
 
   useEffect(() => {
     const cookies = cookie.parse(document.cookie);
@@ -20,46 +50,26 @@ const Lobby = () => {
     if (!socket.connected) {
       socket.connect({ githubId: id });
     }
+    const enterCreatedRoom = (roomId) => {
+      history.push(`/room/${roomId}`);
+    };
+    socket.onCreateRoom(enterCreatedRoom);
   }, []);
 
   return (
-    <LobbyWrapper>
-      <LobbyHeader>
-        <LobbyNickname>hello,{userName}</LobbyNickname>
-        <div>login</div>
-      </LobbyHeader>
-      <LobbyBody>
-        <CreateRoomButton>+ new Room();</CreateRoomButton>
-        <RoomInfoButton enterable>
-          <div>보현님 언능 들어오세요</div>
-          <div>2/20</div>
-        </RoomInfoButton>
-        <RoomInfoButton enterable={false}>
-          <div>형규님 그만 주무세요</div>
-          <div>2/20</div>
-        </RoomInfoButton>
-        <RoomInfoButton enterable>
-          <div>희선님... 말 잘들을게요 충성^^7</div>
-          <div>2/20</div>
-        </RoomInfoButton>
-        <RoomInfoButton enterable>
-          <div>희선님... 말 잘들을게요 충성^^7</div>
-          <div>2/20</div>
-        </RoomInfoButton>
-        <RoomInfoButton enterable>
-          <div>희선님... 말 잘들을게요 충성^^7</div>
-          <div>2/20</div>
-        </RoomInfoButton>
-        <RoomInfoButton enterable>
-          <div>희선님... 말 잘들을게요 충성^^7</div>
-          <div>2/20</div>
-        </RoomInfoButton>
-        <RoomInfoButton enterable>
-          <div>희선님... 말 잘들을게요 충성^^7</div>
-          <div>2/20</div>
-        </RoomInfoButton>
-      </LobbyBody>
-    </LobbyWrapper>
+    <div>
+      <LobbyWrapper>
+        <LobbyHeader>
+          <LobbyNickname>hello, {userName}</LobbyNickname>
+          <GitHubLoginButton />
+        </LobbyHeader>
+        <LobbyBody>
+          <CreateRoomButton onClick={openRoomCreateModal}>+ new Room();</CreateRoomButton>
+          {RoomInfoButtons()}
+        </LobbyBody>
+      </LobbyWrapper>
+      {isModalOpen ? <RoomCreateModal setOpen={setModalOpen} /> : ''}
+    </div>
   );
 };
